more work on making everything ParArrayGeneric
